Merged PR 9607: 26377 Fix issue in RealtimeResolver
OTP will drop the information about constrained transfers like "staySeated" if the RealtimeResolver feature is turned on and you search with ignoreRealtimeUpdates = true.

This PR fixes this issue and also adds some testing.

The code for setting up the test cases is a bit convoluted. Maybe there is an easier way?

Related work items: #26425