Add preStop hook that sleeps 5s
We are seeing some issues where we are resesok-trip is failing during shutdown because it can't connect to otp. At least some of the cases look like a race condition when the pods are being gracefully shut down but are still getting traffic for some very short amount of time.

I found some sources that recommends adding a sleep in the preStop hook to solve this issue. See for example:
https://blog.gruntwork.io/delaying-shutdown-to-wait-for-pod-deletion-propagation-445f779a8304

Related work items: #67491